requirements.txt
wheel
tensorflow
pillow
visigoth
requests
keras
flask
Pygments